Fix: size_t for int

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,4 +1,5 @@
 #include "lists.h"
+
 /**
  * print_listint - prints all the elements of a listint_t
  * @h: head pointer
@@ -7,7 +8,7 @@
 */
 size_t print_listint(const listint_t *h)
 {
-	int nelem = 0;
+	size_t nelem = 0;
 
 	while (h != NULL)
 	{
